requirements.txt
bs4
discord_webhook
pickledb
python-twitter
requests
selenium
selenium-requests
unidecode
